feat(keyvault-secrets): add browser getting-started sample
Extract InteractiveBrowserCredential usage into a dedicated browser
sample under test/public/samples/browser/gettingStartedBrowser.spec.ts.

- Browser sample compiles to TypeScript-only (no JS, requires bundler)
- Node getting-started cleaned up to remove browser credential code
